requirements.txt
Flask==0.7.2
gunicorn==0.12.2
psycopg2
sqlalchemy-migrate